INCOMPATIBLE change to «»-delimited lists
This is a useful, powerful, and natural generalization of and
improvement over the previous meaning of the «»-delimited lists.

Unfortunately, it is incompatible with the previous functionality of
«»-delimited lists.  Any code which used «»-delimited lists
previously (only available if the `include-utf8' option to
`enable-curry-compose-reader-macros' is set) MUST BE CHANGED.

To preserve previous the behavior change all existing instances of
"«...»" to "«list ...»".

See the README for a description of the new behavior.  A join function
is now required as the first element of «»-delimited lists.  Previously
the `list' function was implicitly used to join all «»-delimited lists.